xml/parsing/manakai/eldecls-perefs-1.dat

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT %fuga;EMPTY>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT xy EMPTY>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT %fuga; EMPTY>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT xy EMPTY>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT ab (%fuga; )>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT ab ( %fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga " xy">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy,aa">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy, aa)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy |  aa">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (xy | aa)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "|">
<!ELEMENT ab (bb%fuga;w)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;xml:dtd:cm:entity begins with connector;;;s
[1]2;1;xml:dtd:ext decl;;;w
[1]2;17;xml:dtd:cm:entity ends with connector;;;s
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (bb | w)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "+">
<!ELEMENT ab (foo%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;xml:dtd:no group delimiter;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (foo)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "+">
<!ELEMENT ab foo%fuga;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;string after md def;;;m
[1]2;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:unknown content keyword;;foo;m
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab >
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "(xy |  aa)">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab ((xy | aa))>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "(#PCDATA) ">
<!ELEMENT ab %fuga;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (#PCDATA)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "#PCDATA">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;1;no XML decl;;;s
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (#PCDATA)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "PCDATA">
<!ELEMENT ab (#%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;xml:dtd:no group delimiter;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (#)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "#PCDATA|abc">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;1;no XML decl;;;s
[1]2;14;xml:dtd:cm:bad mixed repetition;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (#PCDATA | abc)*>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "ab">
<!ELEMENT ab (%fuga;cd)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;21;xml:dtd:no group delimiter;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (ab)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "ab">
<!ELEMENT ab (cd%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;xml:dtd:no group delimiter;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (cd)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "DATA">
<!ELEMENT ab (#PC%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;xml:dtd:no group delimiter;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (#PC)>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "TY">
<!ELEMENT ab EMP%fuga;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;18;string after md def;;;m
[1]2;1;xml:dtd:unknown content keyword;;EMP;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab >
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "EMP">
<!ELEMENT ab %fuga;TY>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;20;string after md def;;;m
[1]2;1;xml:dtd:unknown content keyword;;EMP;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab >
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "EMPTY>">
<!ELEMENT ab %fuga;>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;23;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab EMPTY>
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "EMPTY>">
<!ELEMENT ab (%fuga;)>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;23;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT ab (EMPTY)>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT foo (b)%a;>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
[1]1;42;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo (b)>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT foo ((b)%a;)>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
[1]1;42;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo ((b))>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT foo (b)+%a;>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
[1]1;42;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo (b)+>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT foo ((b) %a;)>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
[1]1;42;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo ((b))>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT foo ((b)+%a;)>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
[1]1;42;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo ((b)+)>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT foo ANY%a;>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
[1]1;42;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo ANY>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT foo (%a;)>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
[1]1;42;xml:dtd:ext decl;;;w
[1]1;56;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT foo %a;>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
[1]1;42;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a ">"><!ELEMENT %a;>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;39;mdc in pe in md;;;m
#document
| <!DOCTYPE a "" "http://x/">
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a "abc>"><!ELEMENT %a;>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;42;mdc in pe in md;;;m
[1]1;45;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT abc >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a SYSTEM "http://x/"><a/>
#resource http://x/
<?xml encoding="utf-8"?><!ENTITY % a "(abc)>"><!ELEMENT foo %a;>
#errors
[1]1;25;xml:dtd:ext decl;;;w
[1]1;44;mdc in pe in md;;;m
[1]1;47;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://x/">
|   <!ELEMENT foo (abc)>
| <{} a>
